[MCCAS] Implement Tag-based splitting of DIEs
Using the block types created in the previous commit, this patch changes
the MCCAS layer to split debug information using a TAG based approach.
The implementation is guarded by a flag and materialization is _not_
part of this patch. The first kind of materialization that will be
implemented (in a subsequent commit) is printing with the llvm-cas-dump
tool.

For now, only DIEs whose tag is `DW_TAG_subprogram` and whose attributes
include a `DW_AT_low_pc` cause a new CAS block to be created.