[lldb][SymbolFileDWARF][NFC] Remove unnecessary calls to GetDWARFDeclContext
The function FindDefinitionTypeForDWARFDeclContext loops over all DIEs
corresponding to types with a certain name and compares the context of each
found DIE with the context of a target DIE. However, the target DIE never
changes throughout this search, and yet we recompute its DeclContext on every
iteration of the search. This is wasteful because the method is not exactly free
(see DWARFDebugInfoEntry::GetDWARFDeclContextStatic).