[lldb][swift] Fix CFA/async register unwinding
The CFA of an async frame is defined as the address of its async context, as
such it does not depend on the frame above it, unlike what the current code
implies. This patch fixes the issue by providing a single calculation for the
CFA in the unwind plan of virtual frames:

```
  row->GetCFAValue().SetIsRegisterDereferenced(regnums->async_ctx_regnum);
```

This is correct for the following reasons:

1. A virtual frame V is always below some _real_ async frame on the stack.
2. Because it is below a real async frame, the virtual frame V _must_ be a
continuation funclet (i.e. "await resume" or "Q" funclet).
3. Continuation funclets have an indirect ABI: their async registers point
**not** to their own async context, but to the async context of the frame
immediately above it. As such, V's async register contains the async context of
frame V-1.
4. The first field of the async context of frame V-1 contains a pointer to its
continuation's async context, that is, the async context of frame V.

Similarly, this patch unifies the calculation for the V+1's async register.
Previously, the code branched and provided two methods of calculating it,
implying this calculation depends on what the V-1 frame is, but it doesn't. This
is explained below, but note that these two frames are at a "distance of 2" of
each other. Implying some dependency between their unwind plans sounds fishy.

```
// Method 1:
row->SetRegisterLocationToIsCFAPlusOffset(regnums->async_ctx_regnum, 0, false);

// Method 2:
static const uint8_t async_dwarf_expression_arm64[] = {
    llvm::dwarf::DW_OP_regx, arm64_dwarf::x22,
    llvm::dwarf::DW_OP_deref,
    llvm::dwarf::DW_OP_deref,
};
```

These are exactly the same thing. Here's why:

0. We established before that all virtual frames have indirect ABI.
1. If virtual frame V has indirect ABI, so does frame V+1.
2. It follows that the async register of V+1 must point to V's async context
(Method 1)
3. Because V has indirect ABI, its async register points to V-1 async context.
Dereferencing it once, we obtain V's async context. Dereferencing it one more
time, we obtain V+1's async context.