[lldb][swift] Skip prologue when unwiding virtual frames
Variables are never available in the prologue of a function. By unwinding the pc
of the virtual frames as the first instruction of the continuation function,
LLDB is unable to display any variable information.

This patch addresses the issue by computing the continuation pointer from the
context of the function being unwound, and then querying the continuation's
SymbolContext to determine its prologue size.