[lldb] Flip the conservative anser for StackID::IsCFAOnStack
The optimization in https://github.com/swiftlang/llvm-project/pull/7877
accidentally changed the default answer for this function.

It's important to default to "On Stack" when we don't know, because that's where
most CFAs live, and because we don't want to trigger the expensive heap
comparisons unless we're absolutely sure the CFA is on the heap.