[lldb] Implement SwiftLanguage::AreEquivalentFunctions
This checks whether the two provided symbol contexts refer to funclets of the
same async function.

In particular, this fixes an issue in the ThreadPlanStepOverRange algorithm for
async functions.
When the plan stops, it compares the current symbol context to the symbol
context of where it started. An initial comparison is done through StackIDs. If
that comparison says the two are equal, the plan then attempts to confirm the
two SymbolContexts are the same by checking whether the underlying functions are
the same. Because different funclets correspond to different low level
functions, this comparison was failing.