[lldb] Implement SwiftLanguageRuntime::AreFuncletsOfSameAsyncFunction
This is going to be useful for the Swift Language plugin to identify whether two
symbol contexts are equivalent, an operation which is done by thread plans such
as ThreadPlanStepOverRange.

The implementation compares the following attributes of the mangling tree:

* If these funclets come from a simple async function (as opposed to a closure),
we compare the "Function" child node in the tree.

* If these funclets come from an async closure:
** Their "Number" child node must be the same. This number distinguishes
different closures in the same scope.
** The closure type must be the same.
** If they have a parent closure, it must be identical.
** If they have a parent function, it must be identical.