[lldb][NFC]  Remove 'list' from the name of a map variable
The variable `m_plans_list` is _not_ a list, but a map. This makes it
very confusing to follow the code around it, as the reader has no
indication about the contents of this variable.

This commit renames it to `tid_to_plan_stack`, to its effect of mapping
thread ids to thread plan stacks. It also replaces unordered_map with a
DenseMap, for a more llvm friendly API.