[lldb] Add OS plugin property for reporting all threads
Currently, an LLDB target option controls whether plugins report all
threads. However, it seems natural for this knowledge could come from
the plugin itself. To support this, this commits adds a virtual method
to the plugin base class, returning an `optional<bool>` to denote
whether the plugin has an opinion on whether it reports all threads. The
default answer of `noopt` causes Process to query the Target option,
maintaining compatibility with the old approach.